requirements.txt
tweetstream
python-rpi.gpio
